+++ packages/react-magma-dom/src/components/TreeView/TreeView.stories.tsx
@@ -2976,7 +2976,7 @@ export const ComplexWithAdditionalContent = {
           wrap={FlexWrap.nowrap}
         >
           <Flex behavior={FlexBehavior.item}>
-            <Hyperlink to="google.com" target="_blank" hasUnderline={false}>
+            <Hyperlink to="google.com" opensInNewTab hasUnderline={false}>
               Most common activity length is 39 chars but what if longer
             </Hyperlink>
           </Flex>
@@ -3451,7 +3451,7 @@ export const VirtualizedLargeTree = {
           wrap={FlexWrap.nowrap}
         >
           <Flex behavior={FlexBehavior.item}>
-            <Hyperlink to="google.com" target="_blank" hasUnderline={false}>
+            <Hyperlink to="google.com" opensInNewTab hasUnderline={false}>
               Most common activity length is 39 chars but what if longer
             </Hyperlink>
           </Flex>